requirements.txt
click==8.0.3
Deprecated==1.2.13
docxcompose==1.3.3
docxtpl==0.14.2
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.4
PuLP==2.6.0
PyJWT==2.3.0
python-docx==0.8.11
pytz==2021.3
redis==4.0.2
six==1.16.0
waitress==2.0.0
Werkzeug==2.0.2
wrapt==1.13.3
